+++ unstructured/nlp/patterns.py
@@ -5,7 +5,7 @@ from typing import Final, List
 # ref: https://stackoverflow.com/questions/16699007/
 # regular-expression-to-match-standard-10-digit-phone-number
 US_PHONE_NUMBERS_PATTERN = (
-    r"(?:\+?(\d{1,3}))?[-. (]{0,3}(\d{3})?[-. )]{0,3}(\d{3})[-. ]{0,3}(\d{4})(?: *x(\d+))?\s*$"
+    r"(?:\+?(\d{1,3}))?[-. (]{0,3}(?:(\d{3})[-. )]{0,3})?(\d{3})[-. ]{0,3}(\d{4})(?: *x(\d+))?\s*$"
 )
 US_PHONE_NUMBERS_RE = re.compile(US_PHONE_NUMBERS_PATTERN)